+++ marathon/generator_test.go
@@ -33,6 +33,11 @@ func TestServicesFromMarathon(t *testing.T) {
 								PortMappings: []PortMapping{PortMapping{ContainerPort: 5000, Protocol: "tcp", ServicePort: 42000}},
 							},
 						},
+						Labels: map[string]string{
+							"proxym.domains":            "docker-registry.unit.test,registry.unit.test",
+							"proxym.port.5000.config":   "option forwardfor\noption httpchk",
+							"proxym.port.5000.protocol": "http",
+						},
 						Ports: []int{42000},
 					},
 					App{
@@ -46,6 +51,10 @@ func TestServicesFromMarathon(t *testing.T) {
 									PortMapping{ContainerPort: 8125, Protocol: "udp", ServicePort: 43002},
 								},
 							},
+						},
+						Labels: map[string]string{
+							"proxym.domains":          "graphite.unit.test",
+							"proxym.port.80.protocol": "http",
 						},
 						Ports: []int{43000, 43001, 43002},
 					},
@@ -107,6 +116,7 @@ func TestServicesFromMarathon(t *testing.T) {
 	require.Len(t, services, 6)
 
 	require.Equal(t, "marathon_redis_6379", services[0].Id)
+	require.Equal(t, "", services[0].Config)
 	require.Len(t, services[0].Domains, 0)
 	require.Equal(t, 6379, services[0].Port)
 	require.Equal(t, "tcp", services[0].TransportProtocol)
@@ -118,25 +128,31 @@ func TestServicesFromMarathon(t *testing.T) {
 	require.Equal(t, services[0].Hosts[1].Port, 31003)
 
 	require.Equal(t, services[1].Id, "marathon_registry_5000")
-	require.Len(t, services[1].Domains, 0)
+	require.Equal(t, "option forwardfor\noption httpchk", services[1].Config)
+	require.Len(t, services[1].Domains, 2)
+	require.Contains(t, services[1].Domains, "docker-registry.unit.test")
+	require.Contains(t, services[1].Domains, "registry.unit.test")
 	require.Equal(t, services[1].Port, 5000)
-	require.Equal(t, services[1].TransportProtocol, "tcp")
+	require.Equal(t, services[1].TransportProtocol, "http")
 	require.Equal(t, services[1].ServicePort, 42000)
 	require.Equal(t, services[1].Source, "Marathon")
 	require.Equal(t, services[1].Hosts[0].Ip, "10.10.10.10")
 	require.Equal(t, services[1].Hosts[0].Port, 31002)
 
 	require.Equal(t, services[2].Id, "marathon_graphite-statsd_80")
-	require.Len(t, services[2].Domains, 0)
+	require.Equal(t, "", services[2].Config)
+	require.Len(t, services[2].Domains, 1)
+	require.Contains(t, services[2].Domains, "graphite.unit.test")
 	require.Equal(t, services[2].Port, 80)
-	require.Equal(t, services[2].TransportProtocol, "tcp")
+	require.Equal(t, services[2].TransportProtocol, "http")
 	require.Equal(t, services[2].ServicePort, 43000)
 	require.Equal(t, services[2].Source, "Marathon")
 	require.Equal(t, services[2].Hosts[0].Ip, "10.10.10.11")
 	require.Equal(t, services[2].Hosts[0].Port, 31001)
 
 	require.Equal(t, services[3].Id, "marathon_graphite-statsd_2003")
-	require.Len(t, services[3].Domains, 0)
+	require.Equal(t, "", services[3].Config)
+	require.Len(t, services[3].Domains, 1)
 	require.Equal(t, services[3].Port, 2003)
 	require.Equal(t, services[3].TransportProtocol, "tcp")
 	require.Equal(t, services[3].ServicePort, 43001)
@@ -145,7 +161,8 @@ func TestServicesFromMarathon(t *testing.T) {
 	require.Equal(t, services[3].Hosts[0].Port, 31002)
 
 	require.Equal(t, services[4].Id, "marathon_graphite-statsd_8125")
-	require.Len(t, services[4].Domains, 0)
+	require.Equal(t, "", services[4].Config)
+	require.Len(t, services[4].Domains, 1)
 	require.Equal(t, services[4].Port, 8125)
 	require.Equal(t, services[4].TransportProtocol, "udp")
 	require.Equal(t, services[4].ServicePort, 43002)
@@ -154,6 +171,7 @@ func TestServicesFromMarathon(t *testing.T) {
 	require.Equal(t, services[4].Hosts[0].Port, 31003)
 
 	require.Equal(t, services[5].Id, "marathon_host-networking_8888")
+	require.Equal(t, "", services[5].Config)
 	require.Len(t, services[5].Domains, 0)
 	require.Equal(t, services[5].Port, 8888)
 	require.Equal(t, services[5].TransportProtocol, "tcp")