+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.6.0 (unreleased)
 
+Improvements:
+* [Marathon] Configure protocol, domains and config through labels
+
 Bug Fixes:
 * [Annotation API] proxym does not exit in case the ZK con is lost [#17](https://github.com/wndhydrnt/proxym/issues/17)
 * [Docs] Fix wrong link to `manager.RegisterHttpHandler`